third_party/blink/web_tests/external/wpt/css/css-fonts/parsing/font-feature-settings-computed-expected.txt

This is a testharness.js-based test.
[FAIL] Property font-feature-settings value '"tnum", "hist"'
  assert_equals: expected "\\"hist\\", \\"tnum\\"" but got "\\"tnum\\", \\"hist\\""
[FAIL] Property font-feature-settings value '"dlig", "smcp", "dlig" 0'
  assert_equals: expected "\\"dlig\\" 0, \\"smcp\\"" but got "\\"dlig\\", \\"smcp\\", \\"dlig\\" 0"
Harness: the test ran to completion.
